Create new web submodule.
